partially making home UI
