+++ package.json
@@ -12,57 +12,39 @@
     "lint:fix": "eslint . --fix"
   },
   "dependencies": {
-    "react-hot-loader": "^4.8.3",
-    "@hot-loader/react-dom": "^16.8.6",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/fontawesome-free": "^5.15.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.12",
+    "@ant-design/icons": "^4.3.0",
     "antd": "^4.8.2",
     "dotenv": "^8.2.0",
-    "gatsby": "^2.24.79",
-    "gatsby-image": "^2.4.21",
-    "gatsby-plugin-sharp": "^2.6.43",
-    "gatsby-plugin-styled-components": "^3.3.14",
-    "gatsby-source-contentful": "^3.0.6",
-    "gatsby-source-filesystem": "^2.3.37",
-    "gatsby-transformer-sharp": "^2.5.20",
-    "popper.js": "^1.16.1",
+    "gatsby": "^2.26.1",
+    "gatsby-image": "^2.5.0",
+    "gatsby-plugin-sharp": "^2.8.0",
+    "gatsby-plugin-styled-components": "^3.4.0",
+    "gatsby-source-filesystem": "^2.5.0",
+    "gatsby-transformer-sharp": "^2.6.0",
+    "prop-types": "^15.7.2",
     "react": "^16.14.0",
-    "react-bootstrap": "^1.4.0",
     "react-dom": "^16.14.0",
-    "react-is": "^16.13.1",
-    "styled-components": "^5.2.1",
-    "typescript": "3.7.5"
+    "styled-components": "^5.2.1"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.0.0",
-    "@typescript-eslint/parser": "^4.0.0",
     "babel-eslint": "^10.1.0",
     "babel-plugin-styled-components": "^1.11.1",
-    "eslint": "^7.8.1",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-airbnb-typescript": "^12.0.0",
+    "eslint": "^7.13.0",
+    "eslint-config-airbnb": "^18.2.1",
     "eslint-config-node": "^4.1.0",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-config-react-app": "^6.0.0",
-    "eslint-config-wesbos": "^1.0.1",
-    "eslint-loader": "^2.2.1",
-    "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-html": "^6.1.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-html": "^6.1.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "peerdeps": "^1.1.1",
-    "prettier": "^2.1.1",
-    "xxxxx": "^1.0.3"
+    "prettier": "^2.1.2"
   },
   "main": "gatsby-config.js",
   "keywords": [],
-  "author": ""
+  "author": "intelsnap"
 }
