+++ src/components/Footer.js
@@ -1,14 +1,30 @@
-import { Link } from 'gatsby'
 import React from 'react'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { Row, Col } from 'antd'
-import '../utils/font-awesome'
+import { Link } from 'gatsby'
 import styled from 'styled-components'
-import intel from '../images/intelf.png'
+import { Row, Col } from 'antd'
+import {
+  TwitterOutlined,
+  LinkedinFilled,
+  GithubFilled,
+  SlackSquareFilled,
+  YoutubeFilled,
+  InstagramFilled,
+} from '@ant-design/icons'
+
+const StyledIcon = styled.a`
+  color: white;
+  font-size: 1.4rem;
+  margin: 5px;
+  :link {
+    color: white;
+  }
+  :hover {
+    color: black;
+  }
+`
 
 const Footer = () => (
   <div
-    className="Foot"
     style={{
       width: '100%',
       background: 'grey',
@@ -20,18 +36,6 @@ const Footer = () => (
     <Row>
       <Col span={3} />
       <Col span={3}>
-        <img
-          src={intel}
-          alt="intelf"
-          width="110"
-          height="82"
-          style={{
-            display: 'block',
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            marginTop: '20px',
-          }}
-        />
         <Row>
           <p
             style={{
@@ -41,7 +45,7 @@ const Footer = () => (
             }}
           >
             <Link
-              to="/privacy"
+              to="/privacy-policy"
               activeStyle={{
                 color: 'rgb(24, 144, 255)',
                 borderBottom: '2px',
@@ -108,56 +112,33 @@ const Footer = () => (
         >
           Get In touch
         </h2>
-        <Ico>
-          <a
-            className="icons"
-            href="https://twitter.com/intelsnap_inc"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'twitter']} size="2x" />
-          </a>
-
-          <a
-            className="icons"
-            href="https://myfavcoffeeplace.io"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'linkedin']} size="2x" />
-          </a>
-
-          <a
-            className="icons"
-            href="https://github.com/intelsnap"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'github']} size="2x" />
-          </a>
-          <a
-            className="icons"
-            href="https://intelsnap.slack.com/join/shared_invite/zt-iiqlbfgi-45A58ck5JHG9xsNqP52DIA#/"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'slack']} size="2x" />
-          </a>
-          <a
-            className="icons"
-            href="https://www.youtube.com/channel/UC8xFo5n_dKRYbG7OoM4QVbg/"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'youtube']} size="2x" />
-          </a>
-          <a className="icons" href="support@intelsnap.com" target="blank">
-            <FontAwesomeIcon icon="envelope" size="2x" />
-          </a>
-
-          <a
-            className="icons"
-            href="https://www.instagram.com/intelsnap_media/"
-            target="blank"
-          >
-            <FontAwesomeIcon icon={['fab', 'instagram']} size="2x" />
-          </a>
-        </Ico>
+        <StyledIcon href="https://twitter.com/intelsnap_inc" target="blank">
+          <TwitterOutlined />
+        </StyledIcon>
+        <StyledIcon href="https://myfavcoffeeplace.io" target="blank">
+          <LinkedinFilled />
+        </StyledIcon>
+        <StyledIcon href="https://github.com/intelsnap" target="blank">
+          <GithubFilled />
+        </StyledIcon>
+        <StyledIcon
+          href="https://intelsnap.slack.com/join/shared_invite/zt-iiqlbfgi-45A58ck5JHG9xsNqP52DIA#/"
+          target="blank"
+        >
+          <SlackSquareFilled />
+        </StyledIcon>
+        <StyledIcon
+          href="https://www.youtube.com/channel/UC8xFo5n_dKRYbG7OoM4QVbg/"
+          target="blank"
+        >
+          <YoutubeFilled />
+        </StyledIcon>
+        <StyledIcon
+          href="https://www.instagram.com/intelsnap_media/"
+          target="blank"
+        >
+          <InstagramFilled />
+        </StyledIcon>
         <Row>
           <p>Copyright 2020 Intelsnap Inc. All Rights Reserved.</p>
         </Row>
@@ -165,17 +146,5 @@ const Footer = () => (
     </Row>
   </div>
 )
-
-const Ico = styled.div`
-  .icons {
-    padding: 08px;
-    color: white;
-    font-size: 0.8rem;
-    :hover {
-      color: black;
-      background-color: gray;
-    }
-  }
-`
 
 export default Footer
