+++ src/pages/404.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import Layout from '../components/layout'
+import Layout from '../components/Layout'
 
-export default function error() {
-  return (
-    <Layout>
-      <h1> This is our Error Page !</h1>
-    </Layout>
-  )
-}
+const Error404 = () => (
+  <Layout>
+    <h1> There was an error !</h1>
+  </Layout>
+)
+
+export default Error404
