--- src/utils/font-awesome.js
@@ -1,22 +1,0 @@
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { faCoffee, faEnvelope } from '@fortawesome/free-solid-svg-icons'
-
-import {
-  faTwitter,
-  faLinkedin,
-  faGithub,
-  faSlack,
-  faYoutube,
-  faInstagram,
-} from '@fortawesome/free-brands-svg-icons'
-
-library.add(
-  faCoffee,
-  faTwitter,
-  faLinkedin,
-  faGithub,
-  faSlack,
-  faYoutube,
-  faEnvelope,
-  faInstagram
-)
